+++ src/IDocumentRenderer.ts
@@ -6,6 +6,6 @@ export interface IDocumentRenderer {
 	isEnabled?: () => boolean;
 	getBodyHtml: (raw: string, languageId: string, options?: any) => string;
 	getTitle?: (uri: vscode.Uri) => string;
-	getCssUriStrings?: () => Array<string>;
-	getResource?: (name: string) => IResourceDescriptor;
+	getCssUriStrings?: (uri: vscode.Uri) => Array<string>;
+	getResource?: (name: string, requestingUri: any) => IResourceDescriptor;
 }